requirements.txt
# S2.T3.1 — merged dependencies (core tutor + ASR + data aug + optional MMS + LoRA).
# Install: pip install -r requirements.txt
#
# Notes:
# - torch / transformers / peft are large; needed for mic ASR (MMS), LoRA (`tutor/llm_qlora`),
#   and HF `datasets` workflows. For text-only demo you still install them today, or use a venv
#   and skip heavy features at runtime (no mic, no LoRA).
# - bitsandbytes is Linux-only in this pin (QLoRA 4-bit); use --no_4bit on Windows CPU.

# --- Core (Gradio tutor, Task 1) ---
cryptography>=42.0.0
gradio>=4.0.0
Pillow>=10.0.0
pyttsx3>=2.90
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
langdetect>=1.0.9

# --- Audio / data (generate_data aug, scripts/child_speech_prepare) ---
librosa>=0.10.0
noisereduce>=3.0.0
soundfile>=0.12.0
datasets>=2.16.0
huggingface-hub>=0.20.0

# --- ASR: Whisper-tiny via faster-whisper ---
faster-whisper>=1.0.0

# --- torch stack: MMS-1B (TUTOR_MMS_ASR=1), LoRA, HF training ---
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.26.0
sentencepiece>=0.1.99

# --- LoRA / QLoRA (tutor/llm_qlora) ---
peft>=0.7.0
bitsandbytes>=0.41.0; platform_system != "Windows"
